+++ src/main/java/com/ppp/billing/service/CustomerService.java
@@ -20,9 +20,9 @@ public interface CustomerService {
 	Page<Customer> pagination(int pageNo, int pageSize);
 	Customer findByEmail(String email);
 	List<Customer> findAll();
-	Customer update(CustomerDTO customerDTO, Long id);
+	Customer update(CustomerDTO customerDTO, int id);
 	File downloadFile(String fileName, String fileStoragePath);
-	void delete(long id);
+	void delete(int id);
 
 	
 }
